Get a nice warning, when php-xml-rss is not installed.

+++ package-updates.php
@@ -1,7 +1,7 @@
 <?
 $title = "Recent Package Updates";
 $cvs_author = '$Author: gecko2 $';
-$cvs_date = '$Date: 2010/09/16 19:42:35 $';
+$cvs_date = '$Date: 2013/02/05 23:47:26 $';
 $is_home = 1;
 
 $metatags = '<meta name="description" content="Fink, a distribution of Unix software for Mac OS X and Darwin">
@@ -9,21 +9,27 @@ $metatags = '<meta name="description" content="Fink, a distribution of Unix soft
 ';
 
 include "header.inc";
-include "XML/RSS.php";
 ?>
 
 <h1>Recent Package Updates</h1>
 
 <?
-$rss = new XML_RSS("news/rdf/fink-unstable-no-splitoffs.rdf");
-$rss->parse();
+$incfile = "/usr/share/php/XML/RSS.php";
+if (file_exists($incfile)) {
+        include "$incfile";
 
-$count = 0;
-foreach($rss->getItems() as $item) {
-	$date = $item['dc:date'];
-	$date = preg_replace('|T.*$|', '', $date);
-	echo "<a style=\"text-decoration: none\" href=\"" . htmlentities($item['link']) . "\" name=\"" . urlencode($item['title']) . "\"><span class=\"news-date\">" . $date . ": </span><span class=\"news-headline\" style=\"text-decoration: underline\">" . $item['title'] . "</span></a><br />\n";
-	echo $item['description'];
+	$rss = new XML_RSS("news/rdf/fink-unstable-no-splitoffs.rdf");
+	$rss->parse();
+
+	$count = 0;
+	foreach($rss->getItems() as $item) {
+		$date = $item['dc:date'];
+		$date = preg_replace('|T.*$|', '', $date);
+		echo "<a style=\"text-decoration: none\" href=\"" . htmlentities($item['link']) . "\" name=\"" . urlencode($item['title']) . "\"><span class=\"news-date\">" . $date . ": </span><span class=\"news-headline\" style=\"text-decoration: underline\">" . $item['title'] . "</span></a><br />\n";
+		echo $item['description'];
+	}
+} else {
+        echo "XML/RSS.php is missing. Please install php-xml-rss.<br/>";
 }
 
 include "footer.inc";
